aprgalgebra refactoring

+++ AprgAlgebra/AprgAlgebra/src/Algebra/Constructs/AdditionAndSubtractionOfTermsOverTerms.cpp
@@ -54,13 +54,12 @@ Terms AdditionAndSubtractionOfTermsOverTerms::getLcmOfDenominatorTerms() const {
 }
 
 Terms AdditionAndSubtractionOfTermsOverTerms::getRevisedNumeratorTermsBasedOnLcmOnIndex(
-    unsigned int itemIndex, Terms const& lcmOfDenominatorTerms) const {
+    int itemIndex, Terms const& lcmOfDenominatorTerms) const {
     Terms numeratorTerms;
-    if (itemIndex < m_items.size()) {
+    if (itemIndex < static_cast<int>(m_items.size())) {
         Terms const& multipliers(lcmOfDenominatorTerms);
         Terms monomialMultiplierTerms;
-        Terms nonMonomialMultiplierTerms;
-        segregateMonomialsAndNonMonomials(multipliers, monomialMultiplierTerms, nonMonomialMultiplierTerms);
+        Terms nonMonomialMultiplierTerms;        segregateMonomialsAndNonMonomials(multipliers, monomialMultiplierTerms, nonMonomialMultiplierTerms);
         Monomial monomialMultiplier(getCombinedMonomialMultiplier(monomialMultiplierTerms));
         updateMonomialAndNonMonomialMultipliersBasedOnDenominatorOnIndex(
             itemIndex, monomialMultiplier, nonMonomialMultiplierTerms);
@@ -106,11 +105,10 @@ Monomial AdditionAndSubtractionOfTermsOverTerms::getCombinedMonomialMultiplier(
 }
 
 void AdditionAndSubtractionOfTermsOverTerms::updateMonomialAndNonMonomialMultipliersBasedOnDenominatorOnIndex(
-    unsigned int itemIndex, Monomial& monomialMultiplier, Terms& nonMonomialMultiplierTerms) const {
+    int itemIndex, Monomial& monomialMultiplier, Terms& nonMonomialMultiplierTerms) const {
     for (Term const& denominatorTerm : m_items.at(itemIndex).getDenominators()) {
         if (canBeConvertedToMonomial(denominatorTerm)) {
-            monomialMultiplier.divideMonomial(createMonomialIfPossible(denominatorTerm));
-        } else {
+            monomialMultiplier.divideMonomial(createMonomialIfPossible(denominatorTerm));        } else {
             Terms::iterator multiplierIterator =
                 find(nonMonomialMultiplierTerms.begin(), nonMonomialMultiplierTerms.end(), denominatorTerm);
             if (multiplierIterator != nonMonomialMultiplierTerms.end()) {
@@ -120,12 +118,10 @@ void AdditionAndSubtractionOfTermsOverTerms::updateMonomialAndNonMonomialMultipl
     }
 }
 
-void AdditionAndSubtractionOfTermsOverTerms::emplaceExistingNumeratorTerms(
-    Terms& numeratorTerms, unsigned int itemIndex) const {
+void AdditionAndSubtractionOfTermsOverTerms::emplaceExistingNumeratorTerms(Terms& numeratorTerms, int itemIndex) const {
     for (Term const& numeratorTerm : m_items.at(itemIndex).getNumerators()) {
         numeratorTerms.emplace_back(numeratorTerm);
-    }
-}
+    }}
 
 void AdditionAndSubtractionOfTermsOverTerms::emplaceMonomialMultiplierIfNeeded(
     Terms& numeratorTerms, Monomial const& monomialMultiplier) const {
@@ -144,12 +140,11 @@ void AdditionAndSubtractionOfTermsOverTerms::emplaceNonMonomialMultipliers(
 Expression AdditionAndSubtractionOfTermsOverTerms::getCombinedNumeratorExpression(
     Terms const& lcmDenominatorTerms) const {
     Expression combinedNumerator;
-    unsigned int numberOfItems = m_items.size();
-    for (unsigned int numeratorIndex = 0; numeratorIndex < numberOfItems; numeratorIndex++) {
+    int numberOfItems = m_items.size();
+    for (int numeratorIndex = 0; numeratorIndex < numberOfItems; numeratorIndex++) {
         Expression combinedNumeratorOnIndex(
             getCombinedExpressionForNumeratorOnIndex(numeratorIndex, lcmDenominatorTerms));
-        combineExpressionAsAddOrSubtract(
-            combinedNumerator, combinedNumeratorOnIndex, m_associations.at(numeratorIndex));
+        combineExpressionAsAddOrSubtract(            combinedNumerator, combinedNumeratorOnIndex, m_associations.at(numeratorIndex));
     }
     return combinedNumerator;
 }
@@ -164,11 +159,10 @@ Expression AdditionAndSubtractionOfTermsOverTerms::getCombinedDenominatorExpress
 }
 
 Expression AdditionAndSubtractionOfTermsOverTerms::getCombinedExpressionForNumeratorOnIndex(
-    unsigned int numeratorIndex, Terms const& lcmDenominatorTerms) const {
+    int numeratorIndex, Terms const& lcmDenominatorTerms) const {
     Expression combinedNumeratorOnIndex(Term(1));
     Terms numeratorTermsOnIndex(getRevisedNumeratorTermsBasedOnLcmOnIndex(numeratorIndex, lcmDenominatorTerms));
-    for (Term const& numeratorTermOnIndex : numeratorTermsOnIndex) {
-        combinedNumeratorOnIndex.putTermWithMultiplicationIfNeeded(numeratorTermOnIndex);
+    for (Term const& numeratorTermOnIndex : numeratorTermsOnIndex) {        combinedNumeratorOnIndex.putTermWithMultiplicationIfNeeded(numeratorTermOnIndex);
     }
     return combinedNumeratorOnIndex;
 }
